requirements.txt
streamlit==0.86.0
plotly==5.3.1
Pillow==8.4.0
numpy==1.21.2